# openapi.yaml
openapi: 3.1.0
info:
  title: Spigot API
  version: '1.0.0'
  description: |
    Pay-per-call API marketplace settled in USDC on Arc Testnet.

    **Two ways to pay.** Either works against the same endpoints.

    1. **Credits** — sign in, buy credit once from your wallet, then call any listed
       API with a single authenticated request. The gateway signs and settles each
       payment for you, so there is no wallet popup per call and you never hold gas tokens.
    2. **Direct Nanopayments** — call endpoints with signed nanopayment authorization.
       The gateway / facilitator covers gas, so the caller needs no native gas tokens either.

    **Generating a client.** This document is the source of truth. Any OpenAPI
    generator will produce a client in your language:

    ```bash
    npx @openapitools/openapi-generator-cli generate \
      -i https://raw.githubusercontent.com/0xSarnavo/spigot-arc/main/openapi.yaml \
      -g python -o ./arc-client
    ```
  license:
    name: Proprietary — all rights reserved
  contact:
    name: Spigot
    url: https://github.com/0xSarnavo/spigot-arc
servers:
  - url: http://localhost:4402
    description: Local development
  - url: https://your-deployment.example.com
    description: Your deployment
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/auth/otp:
    post:
      tags: [Auth]
      summary: Email a 6-digit sign-in code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
                  channel:
                    type: string
                    description: '`resend` when email is configured, otherwise `console`'
                  expiresInMinutes: { type: integer }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/auth/otp/verify:
    post:
      tags: [Auth]
      summary: Exchange the code for a session
      description: |
        Sets an httpOnly session cookie. Headless clients (CLI, scripts, CI) should send
        `mode: "token"` or the header `x-client-type: cli` to receive the raw token and
        send it back as `Authorization: Bearer <token>`.

        The token is never returned by default, because handing it to a browser in a
        response body would expose it to any cross-site scripting flaw.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: '^[0-9]{6}$' }
                mode:
                  type: string
                  enum: [token]
                  description: Return the raw session token for headless use
      responses:
        '200':
          description: Signed in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
        '401': { description: Incorrect or expired code }
  /api/auth/magic-link:
    post:
      tags: [Auth]
      summary: Email a single-use sign-in link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200': { description: Link sent }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/auth/google:
    post:
      tags: [Auth]
      summary: Sign in with a Google ID token
      description: The token is verified server-side against Google's public keys.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential:
                  type: string
                  description: ID token from Google Identity Services
      responses:
        '200':
          description: Signed in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
        '401': { description: Invalid credential }
        '503': { description: Google sign-in is not configured on this server }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: x402_session
      description: Set by sign-in. Used by browsers.
    bearerToken:
      type: http
      scheme: bearer
      description: The same session token, for CLI and server-to-server clients.
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
  responses:
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        emailVerified: { type: boolean }
        authProviders:
          type: array
          items: { type: string, enum: [google, magic_link, otp] }
        publisherId: { type: string, nullable: true }
    SessionResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        isNew: { type: boolean }
        token:
          type: string
          description: 'Only present when the caller opted in with mode=token'
        expiresAt: { type: string, format: date-time }
...